requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-jalali==6.0.0
django-jalali-date==1.0.1
django-phone-field==1.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
jdatetime==4.1.0
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
